lamhosts {Rmpi}	R Documentation

Hosts Information

Description

lamhosts finds the host name associated with its node number. Can be used by mpi.spawn.Rslaves to spawn R slaves on selected hosts. This is a LAM-MPI specific function.

mpi.is.master checks if it is running on master or slaves.

mpi.hostinfo finds an individual host information including rank and size in a comm.

slave.hostinfo is executed only by master and find all master and slaves host information in a comm.

mpichhosts finds host names from master Windows registery database.

Usage


lamhosts()
mpi.is.master()
mpi.hostinfo(comm = 1)
slave.hostinfo(comm = 1, short=TRUE) 
mpichhosts()

Arguments

comm
a communicator number

short
if true, a short form is printed

Value

lamhosts returns CPUs nodes numbers with their host names.

mpi.is.master returns TRUE if it is on master and FALSE otherwise.

mpi.hostinfo sends to stdio a host name, rank, size and comm.

slave.hostname sends to stdio a list of host, rank, size, and comm information for all master and slaves. With short=TRUE and 8 slaves or more, the first 3 and last 2 slaves are shown.

mpichhosts returns hostnames including multiple same hostname if a host has multiple CPUs. The first host is labeled as "master", the rest is labeled as "slave1", "slave2", and so on. mpi.universe.size will return the length of mpichhosts-1 on Windows platform.

Author(s)

Hao Yu

See Also

mpi.spawn.Rslaves

[Package Rmpi version 0.6-6 Index]