mpi.bcast {Rmpi}	R Documentation

MPI\_Bcast API

Description

mpi.bcast is a collective call among all members in a comm. It broadcasts a message from the specified rank to all members.

Usage


mpi.bcast(x, type, rank = 0, comm = 1)

Arguments

x
data to be sent or received. Must be the same type among all members.

type
1 for integer, 2 for double, and 3 for character. Others are not supported.

rank
the sender.

comm
a communicator number.

Details

mpi.bcast is a blocking call among all members in a comm, i.e, all members have to wait until everyone calls it. All members have to prepare the same type of messages (buffers). Hence it is relatively difficult to use in R environment since the receivers may not know what types of data to receive, not mention the length of data. Users should use various extensions of mpi.bcast in R. They are mpi.bcast.Robj, mpi.bcast.cmd, and mpi.bcast.Robj2slave.

Value

mpi.bcast returns the message broadcasted by the sender (specified by the rank).

References

http://www.lam-mpi.org/, http://www-unix.mcs.anl.gov/mpi/www/www3/

See Also

mpi.bcast.Robj, mpi.bcast.cmd, mpi.bcast.Robj2slave.

[Package Rmpi version 0.6-6 Index]