mpi.cart.create {Rmpi}	R Documentation

MPI\_Cart\_create

Description

mpi.cart.create creates a Cartesian structure of arbitrary dimension.

Usage


 mpi.cart.create(commold=1, dims, periods, reorder=FALSE, commcart=3)

Arguments

commold
Input communicator

dims
Integery array of size ndims specifying the number of processes in each dimension

periods
Logical array of size ndims specifying whether the grid is periodic or not in each dimension

reorder
ranks may be reordered or not

commcart
The new communicator to which the Cartesian topology information is attached

Details

If reorder = false, then the rank of each process in the new group is the same as its rank in the old group. If the total size of the Cartesian grid is smaller than the size of the group of commold, then some processes are returned mpi.comm.null. The call is erroneous if it specifies a grid that is larger than the group size.

Value

mpi.cart.create returns 1 if success and 0 otherwise.

Author(s)

Alek Hunchak and Hao Yu

References

http://www.lam-mpi.org/, http://www-unix.mcs.anl.gov/mpi/www/www3/

Examples


## Not run: 
#Need at least 9 slaves
mpi.bcast.cmd(mpi.cart.create(1,c(3,3),c(F,T)))
mpi.cart.create(1,c(3,3),c(F,T))

## End(Not run)

[Package Rmpi version 0.6-6 Index]