mpi.cart.get {Rmpi}	R Documentation

MPI\_Cart\_get

Description

mpi.cart.get provides the user with information on the Cartesian topology associated with a comm.

Usage


 mpi.cart.get(comm=3, maxdims)

Arguments

comm
Communicator with Cartesian structure

maxdims
length of vectors dims, periods, and coords in the calling program

Details

The coords are as given for the rank of the calling process as shown.

Value

mpi.cart.get returns a vector containing information on the Cartesian topology associated with comm. maxdims must be at least ndims as returned by mpi.cartdim.get.

Author(s)

Alek Hunchak and Hao Yu

References

http://www.lam-mpi.org/, http://www-unix.mcs.anl.gov/mpi/www/www3/

See Also

mpi.cart.create,mpi.cartdim.get

Examples


## Not run: 
#Need at least 9 slaves
mpi.bcast.cmd(mpi.cart.create(1,c(3,3),c(F,T)))
mpi.cart.create(1,c(3,3),c(F,T))
mpi.remote.exec(mpi.cart.get(3,2))

## End(Not run)

[Package Rmpi version 0.6-6 Index]