mpi.cart.rank {Rmpi}	R Documentation

MPI\_Cart\_rank

Description

mpi.cart.rank translates a Cartesian topology coordinate to its rank.

Usage


 mpi.cart.rank(comm=3, coords)

Arguments

comm
Communicator with Cartesian structure

coords
Specifies the Cartesian coordinates of a process

Details

For a process group with a Cartesian topology, this function translates the logical process coordinates to process ranks as they are used by the point-to-point routines. It is the inverse map of mpi.cart.coords.

Value

mpi.cart.rank returns the rank of the specified process.

Author(s)

Alek Hunchak and Hao Yu

References

http://www.lam-mpi.org/, http://www-unix.mcs.anl.gov/mpi/www/www3/

See Also

mpi.cart.coords

Examples


## Not run: 
#Need at least 9 slaves
mpi.bcast.cmd(mpi.cart.create(1,c(3,3),c(F,T)))
mpi.cart.create(1,c(3,3),c(F,T))
mpi.cart.rank(3,c(1,0))

## End(Not run)

[Package Rmpi version 0.6-6 Index]