mpi.cartdim.get {Rmpi}	R Documentation

MPI\_Cartdim\_get

Description

mpi.cartdim.get gets dim information about a Cartesian topology.

Usage


 mpi.cartdim.get(comm=3)

Arguments

comm
Communicator with Cartesian structure

Details

Can be used to provide other functions with the correct size of arrays.

Value

mpi.cartdim.get returns the number of dimensions of the Cartesian structure

Author(s)

Alek Hunchak and Hao Yu

References

http://www.lam-mpi.org/, http://www-unix.mcs.anl.gov/mpi/www/www3/

See Also

mpi.cart.get

Examples


## Not run: 
#Need at least 9 slaves
mpi.bcast.cmd(mpi.cart.create(1,c(3,3),c(F,T)))
mpi.cart.create(1,c(3,3),c(F,T))
mpi.cartdim.get(comm=3)

## End(Not run)

[Package Rmpi version 0.6-6 Index]