mpi.comm.free {Rmpi}	R Documentation

MPI\_Comm\_free API

Description

mpi.comm.free deallocates a communicator so it points to MPI\_COMM\_NULL.

Usage


  mpi.comm.free(comm=1)

Arguments

comm
a communicator number

Details

When members associated with a communicator finish jobs or exit, they have to call mpi.comm.free to release resource so mpi.comm.size will return 0. If the comm was created from an intercommunicator by mpi.intercomm.merge, use mpi.comm.disconnect instead.

Value

1 if success. Otherwise 0.

Author(s)

Hao Yu

References

http://www.lam-mpi.org/, http://www-unix.mcs.anl.gov/mpi/www/www3/

See Also

mpi.comm.disconnect

[Package Rmpi version 0.6-6 Index]