mpi.comm.spawn {Rmpi}	R Documentation

MPI\_Comm\_spawn API

Description

mpi.comm.spawn tries to start nslaves identical copies of slaves, establishing communication with them and returning an intercommunicator. The spawned slaves are referred to as children, and the process that spawned them is called the parent (master). The children have their own MPI\_COMM\_WORLD represented by comm 0. To make communication possible among master and slaves, all slaves should use mpi.comm.get.parent to find their parent and use mpi.intercomm.merge to merger an intercomm to a comm.

Usage


 mpi.comm.spawn(slave, slavearg = character(0),
                nslaves = mpi.universe.size(), info = 0,
                root = 0, intercomm = 2, quiet = FALSE)

Arguments

slave
a file name to an executable program.

slavearg
an argument list (a char vector) to slave.

nslaves
number of slaves to be spawned.

info
an info number.

root
the root member who spawns slaves.

intercomm
an intercomm number.

quiet
a logical. If TRUE, do not print anything unless an error occurs.

Value

Unless quiet = TRUE, a message is printed to indicate how many slaves are successfully spawned and how many failed.

Author(s)

Hao Yu

References

http://www.lam-mpi.org/, http://www-unix.mcs.anl.gov/mpi/www/www3/

See Also

mpi.comm.get.parent, mpi.intercomm.merge.

[Package Rmpi version 0.6-6 Index]