mpi.dims.create {Rmpi}	R Documentation

MPI\_Dims\_create

Description

mpi.dims.create Create a Cartesian dimension used by mpi.cart.create.

Usage


 mpi.dims.create(nnodes, ndims, dims=integer(ndims))

Arguments

nnodes
Number of nodes in a cluster

ndims
Number of dimension in a Cartesian topology

dims
Initial dimension numbers

Details

The entries in the return value are set to describe a Cartesian grid with ndims dimensions and a total of nnodes nodes. The dimensions are set to be as close to each other as possible, using an appropriate divisibility algorithm. The return value can be constrained by specifying positive number(s) in dims. Only those 0 values in dims are modified by mpi.dims.create.

Value

mpi.dims.create returns the dimension vector used by that in mpi.cart.create.

Author(s)

Hao Yu

References

http://www.lam-mpi.org/, http://www-unix.mcs.anl.gov/mpi/www/www3/

See Also

mpi.cart.create

Examples


#What is the dim numbers of 2 dim Cartersian topology under a grid of 36 nodes
mpi.dims.create(36,2)	#return c(6,6)
#Constrained dim numbers
mpi.dims.create(12,2,c(0,4)) #return c(9,4)

[Package Rmpi version 0.6-6 Index]