mpi.get.count {Rmpi}	R Documentation

MPI\_Get\_count API

Description

mpi.get.count finds the length of a received message.

Usage


mpi.get.count(type, status = 0)

Arguments

type
1 for integer, 2 for double, 3 for char.

status
a status number

Details

When mpi.recv is used to receive a message, the receiver buffer can be set to be bigger than the incoming message. To find the exact length of the received message, mpi.get.count is used to find its exact length. mpi.get.count must be called immediately after calling mpi.recv otherwise the status may be changed.

Value

the length of a received message.

Author(s)

Hao Yu

References

http://www.lam-mpi.org/, http://www-unix.mcs.anl.gov/mpi/www/www3/

See Also

mpi.send, mpi.recv, mpi.get.sourcetag, mpi.probe.

[Package Rmpi version 0.6-6 Index]