mpi.realloc {Rmpi}	R Documentation

Find and increase the lengthes of MPI opaques comm, request, and status

Description

mpi.comm.maxsize, mpi.request.maxsize, and mpi.status.maxsize find the lengthes of comm, request, and status arrayes respectively.

mpi.realloc.comm, mpi.realloc.request and mpi.realloc.status increase the lengthes of comm, request and status arrayes to newmaxsize respectively if newmaxsize is bigger than the original maximum size.

Usage


mpi.realloc.comm(newmaxsize)
mpi.realloc.request(newmaxsize)
mpi.realloc.status(newmaxsize)
mpi.comm.maxsize()
mpi.request.maxsize()
mpi.status.maxsize()

Arguments

newmaxsize
an integer.

Details

When Rmpi is loaded, Rmpi allocs comm array with size 10, request array with 10,000 and status array with 5,000. They should be enough in most cases. They use less than 150KB system memory. In rare case, one can use mpi.realloc.comm, mpi.realloc.request and mpi.realloc.status to increase them to bigger arrayes.

Author(s)

Hao Yu

References

http://www.lam-mpi.org/, http://www-unix.mcs.anl.gov/mpi/www/www3/

[Package Rmpi version 0.6-6 Index]